remove node_env por deprecacao; componetiza dados da empresa
