+++ env_sample
@@ -1,2 +1,1 @@
 REACT_APP_API_URL=http://localhost:8000
-NODE_PATH=src/
